more alerts and css

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 
 class UserController extends Controller
@@ -65,6 +66,10 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        if(Gate::denies('isAdmin')){
+            return redirect()->route('admin.users.index')->with('warning', 'You do not have permission to edit users.');
+        }
+
         $roles = Role::all();
 
         return view('admin.users.edit')->with([
@@ -95,6 +100,10 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('isAdmin')){
+            return redirect()->route('admin.users.index')->with('warning', 'You do not have permission to delete users.');
+        }
+
         $user->roles()->detach();
         $user->delete();
         return redirect()->route('admin.users.index')->with('success', 'User has been deleted.');
